Fix iOS render-grid replay guard fallout
