+++ Packages/Shared/CMUXMobileCore/Sources/CMUXMobileCore/MobileTerminalRenderGridReplay.swift
@@ -54,6 +54,7 @@ public struct MobileTerminalRenderGridReplay: Sendable {
         var bytes = Data()
         let stylesByID = styleMapByID(frame.styles)
         let defaultStyle = stylesByID[0] ?? .default
+        let charsetReset = "\u{0F}\u{1B}(B\u{1B})B\u{1B}*B\u{1B}+B"
         let rowsToClear = Set(frame.clearedRows).union(frame.rowSpans.map(\.row)).sorted()
         for row in rowsToClear {
             bytes.append(sgrBytes(for: defaultStyle))
@@ -91,10 +92,14 @@ public struct MobileTerminalRenderGridReplay: Sendable {
 
         // Apply the whole restore inside a synchronized update so the client
         // never presents the empty reset/clear frame before the snapshot lands.
-        // `ESC c` clears immediately before synchronized output can be enabled,
-        // so use a soft reset plus explicit primary-screen clear instead.
-        bytes.append(Data("\u{1B}[?2026h".utf8))
-        bytes.append(Data("\u{1B}[!p\u{1B}[?2026h\u{1B}[?1049l\u{1B}[H\u{1B}[2J\u{1B}[3J".utf8))
+        // Avoid `ESC c`: RIS clears before synchronized output can be enabled.
+        // These are Ghostty-supported resets for state the replay depends on:
+        // main display, charset mapping, scroll margins, tabs, active screen,
+        // cursor position, viewport contents, and scrollback.
+        bytes.append(Data((
+            "\u{1B}[?2026h\u{1B}[0$}\u{1B}[r\u{1B}[?69l\u{1B}[?5W\u{1B}[?1049l" +
+            charsetReset + "\u{1B}[H\u{1B}[2J\u{1B}[3J"
+        ).utf8))
 
         // Dynamic default colors (OSC 10/11/12). Cells already carry explicit
         // RGB, so these mainly fix the cursor color and color queries.
@@ -121,6 +126,7 @@ public struct MobileTerminalRenderGridReplay: Sendable {
                 terminateLast: true
             )
             bytes.append(Data("\u{1B}[?1049h".utf8))
+            bytes.append(Data(charsetReset.utf8))
             bytes.append(sgrBytes(for: defaultStyle))
             appendFlowLines(
                 &bytes,
